+++ TallyBakery.Application.Core/Entities/PriceBreaks.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TallyBakery.Application.Core.Entities
+{
+    public class PriceBreaks
+    {
+        public int qty { get; set; }
+        public decimal price { get; set; }
+    }
+}
