+++ TallyBakery.Application.Core/Entities/Product.cs
@@ -2,18 +2,12 @@
 using System.Linq;
 using System.Collections.Generic;
 
-namespace TallyBakery.Application.Core
+namespace TallyBakery.Application.Core.Entities
 {
     public class Product
     {
         public string code { get; set; }
         public string name { get; set; }
         public List<PriceBreaks> priceBreaks { get; set; }
-    }
-
-    public class PriceBreaks
-    {
-        public int qty { get; set; }
-        public decimal price { get; set; }
     }
 }
