+++ TallyBakery.Application.Core/Interfaces/ITallyBakery.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TallyBakery.Application.Core.Interfaces
+{
+    public interface ITallyBakery
+    {
+    }
+}
