+++ TallyBakery.Application.Core/Orders/Services/OrderServices.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TallyBakery.Application.Core.Orders.Services
+{
+    public class OrderServices
+    {
+    }
+}
