+++ TallyBakery.Application.Core/Products/Services/ProductServices.cs
@@ -3,40 +3,46 @@ using System.Linq;
 using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
+using System;
 
-namespace TallyBakery.Application.Core
+namespace TallyBakery.Application.Core.Products.Services
 {
-    public class ProductRepository{
+    public class ProductServices
+    {
         public List<Product> ProductCollection { get; set; }
-        public ProductRepository()
+        public ProductServices()
         {
-            
+
         }
 
-        public void Initialization(){
+        public void Initialization()
+        {
             var data = File.ReadAllText("~/db.json");
             this.ProductCollection = JsonConvert.DeserializeObject<List<Product>>(data);
         }
 
-        public void GetCheapestPriceBreaks(string code, int qty){
+        public void GetCheapestPriceBreaks(string code, int qty)
+        {
             var productByCode = this.ProductCollection.FirstOrDefault(f => f.code == code);
-            if(productByCode == null)
+            if (productByCode == null)
                 throw new ArgumentNullException($"{nameof(code)} is not the collection");
 
-            var priceBreaks = this.GetMaxPriceBreakByProduct(code, qty);
+            this.GetMaxPriceBreakByProduct(code, qty);
             //qty % ()
             //productByCode
         }
 
-        public void GetMaxPriceBreakByProduct(string code, int qty){
+        private void GetMaxPriceBreakByProduct(string code, int qty)
+        {
             var productByCode = this.ProductCollection.FirstOrDefault(f => f.code == code);
-            if(productByCode == null)
+            if (productByCode == null)
                 throw new ArgumentNullException($"{nameof(code)} is not the collection");
 
             productByCode.priceBreaks.ToList()
-            .ForEach(f=>{
+            .ForEach(f =>
+            {
                 var qtyMod = qty % f.qty;
-                if(qtyMod < f.qty)
+                if (qtyMod < f.qty)
                 {
                     qty = qtyMod;
                     //store the price bracket
