+++ TallyBakery.Application/bin/Debug/netcoreapp3.1/TallyBakery.Application.Core.pdb
