+++ TallyBakery.Application/bin/Debug/netcoreapp3.1/db.json
@@ -1,0 +1,48 @@
+{
+   "products" : [
+        {
+            "code" : "VS5",
+            "name" : "Vegemite Scroll",
+            "priceBreaks":[{
+                "qty":3,
+                "price":6.99
+            },
+            {
+                "qty":5,
+                "price":8.99
+            }]
+        },
+        {
+            "code" : "MB11",
+            "name" : "Blueberry Muffin",
+            "priceBreaks":[{
+                "qty":2,
+                "price":9.99
+            },
+            {
+                "qty":5,
+                "price":16.95
+            },
+            {
+                "qty":8,
+                "price":24.95
+            }]
+        },
+        {
+            "code" : "CF",
+            "name" : "Croissant",
+            "priceBreaks":[{
+                "qty":3,
+                "price":5.95
+            },
+            {
+                "qty":5,
+                "price":9.95
+            },
+            {
+                "qty":9,
+                "price":16.95
+            }]
+        }
+    ]
+}
